requirements.txt
django
djangorestframework
djangorestframework_simplejwt
PyJWT
environ